backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sse-starlette==2.1.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Document parsing
pdfplumber==0.11.4
python-pptx==1.0.2
python-docx==1.1.2
Pillow==10.4.0

# OCR (optional - may skip if not installed)
# paddleocr==2.9.1
# paddlepaddle==3.0.0

# AI
openai>=1.0.0

# PDF export
reportlab>=4.0.0
# xhtml2pdf (unused, using reportlab instead)

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
bcrypt>=4.0.0
PyJWT>=2.8.0
